# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FraudGraph",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transactions/batch": {
      "post": {
        "summary": "Batch",
        "operationId": "batch_api_transactions_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Batch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Batch": {
        "properties": {
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionInput"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Transactions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "transactions"
        ],
        "title": "Batch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Login": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_.@-]+$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Login"
      },
...
      "TransactionInput": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.:@-]+$",
            "title": "Id"
          },
          "occurred_at": {
            "type": "integer",
            "maximum": 32503680000000.0,
            "minimum": 0.0,
            "title": "Occurred At"
          },
          "sender": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.:@-]+$",
            "title": "Sender"
          },
          "receiver": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_.:@-]+$",
            "title": "Receiver"
          },
          "amount_paise": {
            "type": "integer",
            "maximum": 100000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amount Paise"
          },
          "provenance": {
            "type": "string",
            "enum": [
              "synthetic",
              "external_synthetic",
              "real"
            ],
            "title": "Provenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "occurred_at",
          "sender",
          "receiver",
          "amount_paise",
          "provenance"
        ],
        "title": "TransactionInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}